Categories CRUD

+++ api/categories.go
@@ -3,6 +3,7 @@ package api
 import (
 	"github.com/RobertOchmanek/ebiznes_go/database"
 	"github.com/RobertOchmanek/ebiznes_go/model"
+	"github.com/RobertOchmanek/ebiznes_go/model/rest"
 	"net/http"
 	"github.com/labstack/echo/v4"
 )
@@ -12,6 +13,7 @@ func GetCategories(c echo.Context) error {
 	//Obtain current database connection and fetch categories
 	db := database.DbManager()
 	categories := []model.Category{}
+	//Products in each category are excluded when getting all categories
 	db.Find(&categories)
 
 	return c.JSON(http.StatusOK, categories)
@@ -49,8 +51,8 @@ func UpdateCategory(c echo.Context) error {
 	id := c.Param("id")
 
 	//Bind json from request to object
-	updatedCategory := new(model.Category)
-	c.Bind(updatedCategory)
+	restCategory := new(rest.RestCategory)
+	c.Bind(restCategory)
 
 	//Obtain current database connection and update category by ID
 	db := database.DbManager()
@@ -58,7 +60,7 @@ func UpdateCategory(c echo.Context) error {
     db.Where("id = ?", id).Find(&category)
 
 	//Update and save DB object
-	category.Name = updatedCategory.Name
+	category.Name = restCategory.Name
 	db.Save(&category)
 
 	return c.JSON(http.StatusOK, category)
@@ -72,8 +74,12 @@ func RemoveCategory(c echo.Context) error {
 	//Obtain current database connection and remove category by ID
 	db := database.DbManager()
 	category := model.Category{}
-    db.Where("id = ?", id).Find(&category)
-	db.Delete(&category)
+    db.Where("id = ?", id).Preload("Products").Find(&category)
 
-	return c.JSON(http.StatusOK, category)
+	if len(category.Products) > 0 {
+		return c.JSON(http.StatusMethodNotAllowed, "Category can not be removed unless it does not contain any products")
+	} else {
+		db.Delete(&category)
+		return c.JSON(http.StatusOK, category)
+	}
 }
